User Profile Information Updata With Image

+++ app/Http/Controllers/Admin/ProfileController.php
@@ -41,7 +41,11 @@ class ProfileController extends Controller
             $user->email = $request->email;
             $user->save();
 
-            if (file_exists($oldImage)) {
+            $defaultImages = [
+                '/uploads/avatar.png',
+            ];
+
+            if ($oldImage && !in_array($oldImage, $defaultImages) && file_exists($oldImage)) {
                 unlink($oldImage);
             }
